Changed extentions used templating files
The default design of Swiftlet is to have template/view files have
an .html.php extention. While this is descriptive, it does not
conform to the most IDE's and standards which make use the .phtml
extention in favor of the former.

I renamed the template/view files, and updated the View component
to replect these changes.

In addition to that, I've also taken the liberty of correcting the
robots.txt file. I removed the absolute/missing directories(while
leaving /test and a few others) and changed the Swiftlet indicator
to "ezRPG". This should take care of nasty search engines that try
to index library files.