Removed Example script | Replaced with Login info
Swiftlet started us out with example code, that was removed.
Added a check based on session to see if userID is set, see
exRPG/Models/Session.
Created a "loggedIn" variable to mimic the old "LOGGED_IN" constant in
templates.
Made http://ezrpg/index serve up either index.phtml OR home.phtml based
on if user is logged in or not.
Created a "player" variable for the 'home' view to access Player's stats
and info.