Added a number of functions to comply with Rework
Added $messages but not in use yet

Ported over the generateSigature, CompareSignature, and setMessage from
Rework.2b

Created a validUser and loggedIn function.
validUser can be used for pages that require members to be logged in.
Unauthorized visitors are redirected to index.

loggedIn is used to return a bool if the session key 'userid' is set