Fixed login check, added messaging support, etc
Fixed login check:
The system was checking against $player which, if no password or
username was created, $player never existed, so an unneeded Exception
was created. Changed it to check if $fail or $warn were set.

Fixed issue with 404 being displayed due to index not being capitalized
as stated in issue #4

Added support for the message system and removed the unneeded
header('location') since we could just load up a view without doing
another http request.