Added an RBAC [see #13]
It's a basic Flat-RBAC(NIST Level 1), but will suffice for all our
needs. It does not support hierarchies at all, and review still
needs to be added to the database.

I've included a possible new database structure with player layouts
and the settings one(#11) as well. Of course, the AC's tables have
also been created.

Another niftly little thing I've added to the router is partial
matching. THIS IS VERY EXPENSIVE, but it's quite good from a UI
perspective and will allow misspelled words, as well as serve as
a case-insensitive solution for routes.