This will probably be temp: Fixed issue I was having
I was having a strange issue with the current setup in the create
method:
REmoved $value = $this->quote and replaced with $item = $this->quote as
$value is not even declared.
Removed quotes for columns and table name as I never really heard of
using them and it was genereating issues with the prepared statement.
Lastly changed from the PDO BindParms to directly creating the Query
like the old fashion way to give way for debugging and also because the
other way had issues with using Implodes instead of a variable and Too
many tokens to parameter ratio error.
Perhaps this is only an issue for me?