Improved the registration proccess
We were having an issue where the registration process was failing
due to the salt not being generated correctly. I suspect this may
be a NT bug only, since I could not replicate it on GNU/Linux.

I have modified the registration module to include reasons why the
registration failed. This is VERY primitave at the moment, and can
be improved upon(please do!)