Implementation of Cache and ACL
Along with this, I've added support for multiple error types,
these are access-denied and file-not-found.

Additionally, I've also added a configuration option "showExceptions"
which will toggle what the app displays when an exception occurs.

Lastly, I've added changes that will aim to resolve conficts with
booher's latest updates for the installer.