Firmer integration of the routes and the AC [see #13]
I removed types from permissions in favor of the route-level access
paramater. This allows us to define a closer relationship between
the two without sacraficing functionality i.t.o. extendability.

This change implements a security risk into the engine, which causes
routes to default to "all access". This means routes without this
additional paramter defined will not have any permission or role
attached to it. This may change to "members"/"registered player"
in the future - but it's no of much improtance at this time.