Removing dependancies caused by installer [see #12]
Moved the default configuration to a new file "config.install.php".
This removes the need to have it within the public receiver.

Furthermore, I've also moved a lot of the functionality App did
to seperate places.

The Cache will now also attach itself to the container if needed,
removing the need to explicityly check for it's existence.

A few minor tweaks were also made to the installer itself, which
will now not create configuration files on the first call.